+++ src/components/atoms/PageTitle.tsx
@@ -1,6 +1,8 @@
 import React from 'react'
 import Image from 'next/image'
 
+import seongland from '~/public/seongland.png'
+
 export const PageTitle: React.FC<{ title: string }> = ({ title }) => {
   return (
     <div className="flex absolute top-15vh" font="900">
@@ -8,7 +10,10 @@ export const PageTitle: React.FC<{ title: string }> = ({ title }) => {
         <Image
           fill
           alt={title}
-          src="/seongland.svg"
+          src={seongland.src}
+          placeholder="blur"
+          loading="lazy"
+          blurDataURL={seongland.blurDataURL}
           sizes="(max-width: 768px) 50vw,
               (max-width: 1200px) 25vw,
               20vw"
