+++ src/pages/index.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
-import type { NextPage } from 'next'
 
 import dynamic from 'next/dynamic'
 
+import { dynamicImgCards } from '~/scripts/cards'
 import { TypeTitle } from '@/components/atoms/TypeTitle'
 import { PageTitle } from '@/components/atoms/PageTitle'
-import { Applications } from '@/components/organisms/Applications'
-import { Publication } from '@/components/organisms/Publication'
-import { WebPages } from '@/components/organisms/WebPages'
+import { Cards } from '@/components/organisms/Cards'
+
+import type { Card } from '~/scripts/cards'
 
 export const CenterPage = dynamic(async () => import('@/components/atoms/CenterPage'))
 export const ScrollPage = dynamic(async () => import('@/components/templates/ScrollPage'))
@@ -15,7 +15,11 @@ export const ScrollPage = dynamic(async () => import('@/components/templates/Scr
 const HEIGHT = 7
 const PAGES = 4
 
-const Index: NextPage = () => {
+const Index: React.FC<{ applications: Card[]; publications: Card[]; webapps: Card[] }> = ({
+  applications,
+  publications,
+  webapps,
+}) => {
   return (
     <ScrollPage height={HEIGHT}>
       <CenterPage page={1} pages={PAGES}>
@@ -23,18 +27,24 @@ const Index: NextPage = () => {
       </CenterPage>
       <CenterPage page={2} pages={PAGES}>
         <PageTitle title="Web Pages" />
-        <WebPages />
+        <Cards cards={webapps} />
       </CenterPage>
       <CenterPage page={3} pages={PAGES}>
         <PageTitle title="Publication" />
-        <Publication />
+        <Cards cards={publications} />
       </CenterPage>
       <CenterPage page={4} pages={PAGES}>
         <PageTitle title="Applications" />
-        <Applications />
+        <Cards cards={applications} />
       </CenterPage>
     </ScrollPage>
   )
+}
+
+export async function getStaticProps() {
+  return {
+    props: await dynamicImgCards(),
+  }
 }
 
 export default Index
